LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/window-resize-aborts-transition-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_true: Transition must be finished by the window resize expected true got false

Harness Error (FAIL), message = Unhandled rejection: assert_true: Transition must be finished by the window resize expected true got false

TIMEOUT View transitions: Resizing viewport skips the transition Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_true: Transition must be finished by the window resize expected true got false

TIMEOUT View transitions: Resizing viewport skips the transition Test timed out
